[update] ability to add login to an app

+++ actions/common.js
@@ -1,33 +1,49 @@
 const fs = require("fs");
 const path = require("path");
 
-function replaceFile(name, mode, text){
+
+async function replaceAndSave(name, mode, text, after){
 	name = "./sources/"+name;
 
 	let str = fs.readFileSync(name).toString("utf8");
-	const pos = str.indexOf("/*wjet::"+mode+"*/");
+	let pos = str.indexOf(mode);
 	if (pos != -1){
-		str = str.substr(0, pos) + text + "\n"+str.substr(pos);
-		fs.writeFileSync(name, str);
+		const exists = str.indexOf(text) != -1;
+		if (!exists){
+			if (after)
+				pos += mode.length;
+
+			str = str.substr(0, pos) + text + "\n"+str.substr(pos);
+			fs.writeFileSync(name, str);
+		}
 	}
+}
+async function addMarker(name, mode, text, after){
+	mode = "/*wjet::"+mode+"*/";
+	return replaceAndSave(name, mode, text, after);
 }
 
 async function addPlugin(file, name){
-	addImport(file, "plugins", "webix-jet");
+	addImport(file, "{plugins}", "webix-jet");
 
 	const plugin = `this.use(plugins.${name}, { model: session });`;
-	return replaceFile(file, "plugin", plugin);
+	return addMarker(file, "plugin", plugin);
 }
 
 async function addImport(file, name, from){
+	file =  "./sources/" + file;
 	const line = `import ${name} from "${from}";\n`;
 
-	let str = fs.readFileSync(name).toString("utf8");
+	let str = fs.readFileSync(file).toString("utf8");
 	const pos = str.indexOf(line);
-	if (pos != -1){
+	if (pos === -1){
 		str = line + str;
-		fs.writeFileSync(name, str);
+		fs.writeFileSync(file, str);
 	}
+}
+
+async function addUI(file, code){
+	replaceAndSave(file, "config(){", "\n"+code+"\n\n", true);
 }
 
 async function addSettings(){
@@ -61,5 +77,7 @@ module.exports = {
 	addView,
 	addFile,
 	addModel,
-	addImport
+	addImport,
+	addUI,
+	addMarker
 }
